Finish

+++ classes/Ijdb/Controllers/Joke.php
@@ -63,6 +63,15 @@ class Joke
 
     public function delete()
     {
+
+        $author = $this->authentication->getUser();
+
+        $joke = $this->jokesTable->findById($_POST['id']);
+
+        if ($joke['authorId'] != $author['id']) {
+            return;
+        }
+
         $this->jokesTable->delete($_POST['id']);
 
         header('location: /joke/list');
@@ -70,6 +79,14 @@ class Joke
     public function saveEdit()
     {
         $author = $this->authentication->getUser();
+
+        if (isset($_GET['id'])) {
+            $joke = $this->jokesTable->findById($_GET['id']);
+
+            if ($joke['authorId'] != $author['id']) {
+                return;
+            }
+        }
 
         $joke = $_POST['joke'];
         $joke['jokedate'] = new \DateTime();
@@ -82,6 +99,8 @@ class Joke
 
     public function edit()
     {
+        $author = $this->authentication->getUser();
+
         if (isset($_GET['id'])) {
             $joke = $this->jokesTable->findById($_GET['id']);
         }
@@ -92,7 +111,8 @@ class Joke
             'template' => 'editjoke.html.php',
             'title' => $title,
             'variables' => [
-                'joke' => $joke ?? null
+                'joke' => $joke ?? null,
+                'userId' => $author['id'] ?? null
             ]
         ];
     }
